+++ package.json
@@ -8,6 +8,7 @@
     "@testing-library/user-event": "^13.5.0",
     "axios": "^1.4.0",
     "cheerio": "^1.0.0-rc.12",
+    "cors": "^2.8.5",
     "detergent": "^9.2.5",
     "express": "^4.18.2",
     "react": "^18.2.0",
